Create check_parentheses_and_brackets.c
Follow a given sequence of parentheses, braces, and square brackets stored in a string s. Verify if the sequence   is well-formed. It means brackets, braces, and square brackets are closed in the reverse order of their opening. To see the result input a value equal to “###”.

Example:

Input:

(({[{()}([])]}))
[({)}]
(()))
###


Output:

SIM
NAO
NAO